lyco-langgraph/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langserve[all]>=0.3.0

# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.9

# Database and Caching
redis>=5.0.0
asyncpg>=0.30.0
supabase>=2.9.0
sqlalchemy>=2.0.0

# AI/ML
anthropic>=0.39.0
openai>=1.54.0
tiktoken>=0.8.0

# Data Processing
pydantic>=2.9.0
numpy>=1.26.0
pandas>=2.2.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=24.0.0
httpx>=0.27.0
email-validator>=2.2.0
python-dateutil>=2.9.0

# Monitoring and Debugging
structlog>=24.0.0
prometheus-client>=0.21.0
psutil>=6.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Development
black>=24.0.0
ruff>=0.7.0
mypy>=1.13.0
ipython>=8.29.0
